dom/media/mediasource/test/test_SeekableAfterEndOfStream.html

<!DOCTYPE HTML>
<html>
<head>
  <title>MSE: seekable attribute after end of stream</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="mediasource.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();

runWithMSE(function (ms, v) {
  ms.addEventListener("sourceopen", function () {
    var sb = ms.addSourceBuffer("video/webm");

    fetchWithXHR("seek.webm", function (arrayBuffer) {
      sb.appendBuffer(new Uint8Array(arrayBuffer));
      var promises = [];
      promises.push(once(sb, "updateend"));
      promises.push(once(v, "loadedmetadata"));
      Promise.all(promises).then(function () {
        ms.endOfStream();
        once(ms, "sourceended").then(function() {
          var target = 2;
          ok(v.seekable.length, "Resource is seekable");
          ok(v.seekable.length &&
             target >= v.seekable.start(0) &&
             target < v.seekable.end(0), "Target is within seekable range");
          SimpleTest.finish();
        });
      });
    });
  });
});
</script>
</pre>
</body>
</html>
